Refactor codebase for improved maintainability and transparency.
