+++ src/components/leaflet-map/leaflet-map.tsx
@@ -50,7 +50,7 @@ function LeafletMap(props: Readonly<MapProps>) {
     }
   }, [map, points, selectedPoint]);
 
-  return <div style={{height: '500px'}} ref={mapRef}></div>;
+  return <div style={{height: '100%'}} ref={mapRef}></div>;
 }
 
 export default LeafletMap;
