--- src/mocks/poins.ts
@@ -1,30 +1,0 @@
-export const MAP_CITY = {
-  title: 'Нью-Йорк',
-  lat: 40.835292,
-  lng: -73.916236,
-  zoom: 10,
-};
-
-export const MAP_POINTS = [
-  {
-    title: 'Саундвью',
-    lat: 40.816881,
-    lng: -73.872768,
-  }, {
-    title: 'Ферри Поинт',
-    lat: 40.814909,
-    lng: -73.830682,
-  }, {
-    title: 'Бронкс',
-    lat: 40.862413,
-    lng: -73.879357,
-  }, {
-    title: 'Инвуд-Хилл',
-    lat: 40.870817,
-    lng: -73.927112,
-  }, {
-    title: 'Пелхэм-Бей-Парк',
-    lat: 40.877312,
-    lng: -73.807182,
-  },
-];
